+++ tests/test_inspect.py
@@ -410,6 +410,67 @@ def test_inspect_update_existing_manifest(
     assert a == b
 
 
+def test_inspect_update_existing_ref(
+    rc: RawConfig,
+    cli: SpintaCliRunner,
+    tmpdir: Path,
+    sqlite: Sqlite,
+):
+    # Prepare source data.
+    sqlite.init({
+        'COUNTRY': [
+            sa.Column('ID', sa.Integer, primary_key=True),
+            sa.Column('NAME', sa.Text),
+        ],
+        'CITY': [
+            sa.Column('ID', sa.Integer, primary_key=True),
+            sa.Column('NAME', sa.Text),
+            sa.Column('COUNTRY', sa.Integer, sa.ForeignKey("COUNTRY.ID")),
+        ],
+    })
+
+    # Configure Spinta.
+    rc = configure(rc, sqlite, tmpdir / 'manifest.csv', f'''
+    d | r | m | property | type    | ref | source  | prepare | access  | title
+    datasets/gov/example |         |     |         |         |         | Example
+      | schema           | sql     | sql |         |         |         | 
+                         |         |     |         |         |         |
+      |   | Country      |         | id  | COUNTRY |         |         | Country
+      |   |   | id       | integer |     | ID      |         | private | Primary key
+      |   |   | name     | string  |     | NAME    |         | open    | Country name
+                         |         |     |         |         |         |
+      |   | City         |         | id  | CITY    | id > 1  |         | City
+      |   |   | id       | integer |     | ID      |         | private | 
+      |   |   | name     | string  |     | NAME    | strip() | open    | City name
+      |   |   | country  | integer |     | COUNTRY |         | open    | Country id
+    ''')
+
+    cli.invoke(rc, [
+        'inspect',
+        tmpdir / 'manifest.csv',
+        '-o', tmpdir / 'result.csv',
+        ])
+
+    # Check what was detected.
+    manifest = load_manifest(rc, tmpdir / 'result.csv')
+    a, b = compare_manifest(manifest, '''
+    d | r | b | m | property | type    | ref     | source  | prepare | access  | title
+    datasets/gov/example     |         |         |         |         |         | Example
+      | schema               | sql     | sql     |         |         |         |
+                             |         |         |         |         |         |
+      |   |   | Country      |         | id      | COUNTRY |         |         | Country
+      |   |   |   | id       | integer |         | ID      |         | private | Primary key
+      |   |   |   | name     | string  |         | NAME    |         | open    | Country name
+                             |         |         |         |         |         |
+      |   |   | City         |         | id      | CITY    | id>1    |         | City
+      |   |   |   | id       | integer |         | ID      |         | private |
+      |   |   |   | name     | string  |         | NAME    | strip() | open    | City name
+      |   |   |   | country  | ref     | Country | COUNTRY |         | open    | Country id
+    ''')
+    assert a == b
+
+
+
 def test_inspect_with_empty_config_dir(
     rc: RawConfig,
     cli: SpintaCliRunner,